requirements.txt
streamlit
gitpython
nbformat
groq
python-dotenv
watchdog
fpdf